Fixed incorrect linebreak
The missing line is prevState = 'a' in the lineBreak(). When it was missing, when the word started a new line, it assumed that the prevState was the state of the last character that triggered the line break. If that character was "nf" for example, the start of the word will assume that prevState was "f" and it caused incorrect kerning.

Test case:

sehivg mevbvtu meqeluwrfknf pnsf tubuhf yqf beflebihf . hlf aini beflvsuqf ketik pin befad pd posisi diymf , dnf gjhf dptf meqeluwrfknf lebihf bxkf

Try to adjust the window until there's a lineBreak right before "hlf" and notice that "dnf" is broken incorrectly.  The discrepancy is visible during the initial parse (before adding lineBreaks) where you can see that the "h" was positioned incorrectly.